Move faceIDCS to util.h
