+++ src/Mesh.cpp
@@ -1,6 +1,6 @@
 #include "Mesh.h"
 #include<glm/glm.hpp>
-
+#include"util.h"
 #include <iostream>
 #include <fstream>
 #include <sstream>
@@ -67,75 +67,6 @@ void Model::destroy(){
 //
 //
 /// ----------------
-
-struct FaceIdcs
-{
-	int v[4];
-	int vn[4];
-	int vt[4];
-    bool is_quad;
-
-	FaceIdcs()
-	{
-		for (int i = 0; i < 4; i++)
-			v[i] = vn[i] = vt[i] = 0;
-	}
-
-	FaceIdcs(std::istream &aStream)
-	{
-		for (int i = 0; i < 4; i++)
-			v[i] = vn[i] = vt[i] = 0;
-
-		char c;
-        is_quad = false;
-		for (int i = 0; i < 4; i++)
-		{
-            aStream >> std::ws;
-            if(aStream.peek() == EOF){
-                break;
-            }
-            if(i == 3){
-                is_quad = true;
-            }
-			aStream >> v[i] >> std::ws;
-			if (aStream.peek() != '/')
-				continue;
-			aStream >> c >> std::ws;
-			if (aStream.peek() == '/')
-			{
-				aStream >> c >> std::ws >> vn[i];
-				continue;
-			}
-			else
-				aStream >> vt[i];
-			if (aStream.peek() != '/')
-				continue;
-			aStream >> c >> vn[i];
-		}
-	}
-friend std::ostream& operator<<(std::ostream& os, const FaceIdcs& face) {
-    for (int i = 0; i < 4; i++) {
-        os << "v" << i + 1 << ": " << face.v[i] << ", ";
-        os << "vn" << i + 1 << ": " << face.vn[i] << ", ";
-        os << "vt" << i + 1 << ": " << face.vt[i] << "; ";
-    }
-    return os;
-}
-};
-
-vec3 vec3fFromStream(std::istream &aStream)
-{
-	float x, y, z;
-	aStream >> x >> std::ws >> y >> std::ws >> z;
-	return vec3(x, y, z);
-}
-
-vec2 vec2fFromStream(std::istream &aStream)
-{
-	float x, y;
-	aStream >> x >> std::ws >> y;
-	return vec2(x, y);
-}
 
 void Model::loadFile(const std::string& file){
     std::ifstream ifile;
